chore: renamed SongUtils.java to MusicServiceUtils.java and moved getSong to MusicServiceUtils
