+++ app/src/main/java/com/simplecity/amp_library/ui/screens/album/menu/AlbumMenuPresenter.kt
@@ -1,12 +1,10 @@
 package com.simplecity.amp_library.ui.screens.album.menu
 
-import com.simplecity.amp_library.ShuttleApplication
 import com.simplecity.amp_library.data.Repository
 import com.simplecity.amp_library.model.Album
 import com.simplecity.amp_library.model.Playlist
 import com.simplecity.amp_library.model.Song
 import com.simplecity.amp_library.playback.MediaManager
-import com.simplecity.amp_library.playback.QueueManager
 import com.simplecity.amp_library.ui.common.Presenter
 import com.simplecity.amp_library.ui.screens.album.menu.AlbumMenuContract.View
 import com.simplecity.amp_library.ui.screens.drawer.NavigationEventRelay
@@ -20,15 +18,11 @@ import com.simplecity.amp_library.utils.playlists.PlaylistManager
 import edu.usf.sas.pal.muser.model.UiEventType
 import edu.usf.sas.pal.muser.util.EventUtils
 import edu.usf.sas.pal.muser.util.FirebaseIOUtils
-import edu.usf.sas.pal.muser.util.SongUtils
 import io.reactivex.Observable
 import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
-import java.util.*
-import java.util.logging.Handler
 import javax.inject.Inject
-import kotlin.concurrent.schedule
 
 class AlbumMenuPresenter @Inject constructor(
     private val playlistManager: PlaylistManager,
